// packages/react/compiler/components.generated.ts
export const AirPlayButton = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'AirPlayButton',
  },
});

export const AudioTrackRadioGroup = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'AudioTrackRadioGroup',
  },
});

export const BufferingIndicator = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'BufferingIndicator',
  },
});

export const CaptionsButton = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'CaptionsButton',
  },
});

export const CaptionsRadioGroup = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'CaptionsRadioGroup',
  },
});

export const CastButton = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'CastButton',
  },
});

export const Container = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'Container',
  },
});

export const Controls = {
  Group: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Controls',
      path: ['Group'],
    },
  }),
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Controls',
      path: ['Root'],
    },
  }),
} as const;

export const ErrorDialog = {
  Close: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'ErrorDialog',
      path: ['Close'],
    },
  }),
  Description: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'ErrorDialog',
      path: ['Description'],
    },
  }),
  Popup: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'ErrorDialog',
      path: ['Popup'],
    },
  }),
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'ErrorDialog',
      path: ['Root'],
    },
  }),
  Title: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'ErrorDialog',
      path: ['Title'],
    },
  }),
} as const;

export const FullscreenButton = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'FullscreenButton',
  },
});

export const Gesture = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'Gesture',
  },
});

export const Hotkey = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'Hotkey',
  },
});

export const Menu = {
  CheckboxItem: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['CheckboxItem'],
    },
  }),
  Content: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['Content'],
    },
  }),
  Group: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['Group'],
    },
  }),
  GroupLabel: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['GroupLabel'],
    },
  }),
  Item: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['Item'],
    },
  }),
  ItemIndicator: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['ItemIndicator'],
    },
  }),
  RadioGroup: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['RadioGroup'],
    },
  }),
  RadioItem: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['RadioItem'],
    },
  }),
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['Root'],
    },
  }),
  Separator: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['Separator'],
    },
  }),
  Trigger: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Menu',
      path: ['Trigger'],
    },
  }),
} as const;

export const MuteButton = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'MuteButton',
  },
});

export const PiPButton = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'PiPButton',
  },
});

export const PlaybackRateRadioGroup = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'PlaybackRateRadioGroup',
  },
});

export const PlayButton = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'PlayButton',
  },
});

export const Popover = {
  Arrow: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Popover',
      path: ['Arrow'],
    },
  }),
  Popup: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Popover',
      path: ['Popup'],
    },
  }),
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Popover',
      path: ['Root'],
    },
  }),
  Trigger: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Popover',
      path: ['Trigger'],
    },
  }),
} as const;

export const Poster = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'Poster',
  },
});

export const QualityRadioGroup = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'QualityRadioGroup',
  },
});

export const SeekButton = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'SeekButton',
  },
});

export const SeekIndicator = {
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'SeekIndicator',
      path: ['Root'],
    },
  }),
  Value: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'SeekIndicator',
      path: ['Value'],
    },
  }),
} as const;

export const Slider = {
  Buffer: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Slider',
      path: ['Buffer'],
    },
  }),
  Fill: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Slider',
      path: ['Fill'],
    },
  }),
  Preview: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Slider',
      path: ['Preview'],
    },
  }),
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Slider',
      path: ['Root'],
    },
  }),
  Thumb: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Slider',
      path: ['Thumb'],
    },
  }),
  Thumbnail: {
    Image: defineTarget({
      import: {
        from: '@videojs/react',
        name: 'Slider',
        path: ['Thumbnail', 'Image'],
      },
    }),
    Root: defineTarget({
      import: {
        from: '@videojs/react',
        name: 'Slider',
        path: ['Thumbnail', 'Root'],
      },
    }),
  } as const,
  Track: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Slider',
      path: ['Track'],
    },
  }),
  Value: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Slider',
      path: ['Value'],
    },
  }),
} as const;

export const StatusAnnouncer = defineTarget({
  import: {
    from: '@videojs/react',
    name: 'StatusAnnouncer',
  },
});

export const StatusIndicator = {
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'StatusIndicator',
      path: ['Root'],
    },
  }),
  Value: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'StatusIndicator',
      path: ['Value'],
    },
  }),
} as const;

export const Time = {
  Group: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Time',
      path: ['Group'],
    },
  }),
  Separator: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Time',
      path: ['Separator'],
    },
  }),
  Value: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Time',
      path: ['Value'],
    },
  }),
} as const;

export const TimeSlider = {
  Buffer: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'TimeSlider',
      path: ['Buffer'],
    },
  }),
  Chapters: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'TimeSlider',
      path: ['Chapters'],
    },
  }),
  ChapterTitle: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'TimeSlider',
      path: ['ChapterTitle'],
    },
  }),
  Fill: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'TimeSlider',
      path: ['Fill'],
    },
  }),
  Preview: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'TimeSlider',
      path: ['Preview'],
    },
  }),
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'TimeSlider',
      path: ['Root'],
    },
  }),
  Thumb: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'TimeSlider',
      path: ['Thumb'],
    },
  }),
  Track: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'TimeSlider',
      path: ['Track'],
    },
  }),
  Value: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'TimeSlider',
      path: ['Value'],
    },
  }),
} as const;

export const Tooltip = {
  Arrow: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Tooltip',
      path: ['Arrow'],
    },
  }),
  Label: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Tooltip',
      path: ['Label'],
    },
  }),
  Popup: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Tooltip',
      path: ['Popup'],
    },
  }),
  Provider: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Tooltip',
      path: ['Provider'],
    },
  }),
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Tooltip',
      path: ['Root'],
    },
  }),
  Shortcut: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Tooltip',
      path: ['Shortcut'],
    },
  }),
  Trigger: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'Tooltip',
      path: ['Trigger'],
    },
  }),
} as const;

export const VolumeIndicator = {
  Fill: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'VolumeIndicator',
      path: ['Fill'],
    },
  }),
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'VolumeIndicator',
      path: ['Root'],
    },
  }),
  Value: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'VolumeIndicator',
      path: ['Value'],
    },
  }),
} as const;

export const VolumeSlider = {
  Fill: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'VolumeSlider',
      path: ['Fill'],
    },
  }),
  Preview: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'VolumeSlider',
      path: ['Preview'],
    },
  }),
  Root: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'VolumeSlider',
      path: ['Root'],
    },
  }),
  Thumb: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'VolumeSlider',
      path: ['Thumb'],
    },
  }),
  Track: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'VolumeSlider',
      path: ['Track'],
    },
  }),
  Value: defineTarget({
    import: {
      from: '@videojs/react',
      name: 'VolumeSlider',
      path: ['Value'],
    },
  }),
} as const;

export const targets = {
  AirPlayButton,
  AudioTrackRadioGroup,
  BufferingIndicator,
  CaptionsButton,
  CaptionsRadioGroup,
  CastButton,
  Container,
  Controls,
  ErrorDialog,
  FullscreenButton,
  Gesture,
  Hotkey,
  Menu,
  MuteButton,
  PiPButton,
  PlaybackRateRadioGroup,
  PlayButton,
  Popover,
  Poster,
  QualityRadioGroup,
  SeekButton,
  SeekIndicator,
  Slider,
  StatusAnnouncer,
  StatusIndicator,
  Time,
  TimeSlider,
  Tooltip,
  VolumeIndicator,
  VolumeSlider,
} as const;